The clextree package. 'c' for compressed hmms.
The clextree linguist works similarly to the
'lextree.LexTreeLinguist'. However, instead of maintaining CI units in
the lextree, the clextree version keeps HMMs in the lextree. The
purpose of this is two fold:

1) More of the search space is prebuilt, leaving less work to be done
during decoding (of course this yields a larger memory footprint and
more startup time)

2) More importantly, many of the hmms are shared by different CD
units. The lextree can efficiently represent this sharing, eliminating
the need for separate search states and tokens for units that share
HMMs. This can lead to a significant reduction in beam size.

Some early tests show that for rm1_bigram we can reduce the beam by
half and achieve the same WER with this new linguist. This leads to a
speed improvement of 3X.  More testing will show whether we can expect
this kind of improvement across the board.

Note that this package is woefully under javadoc'd. I normally would
not check it in in this condition, but I want to get this checked in
as quickly as possible to allow ripley folks to tune and try things
out.