Moved functions from sphinx_train.cfg to utils.pl. sphinx_train.cfg now has only configuration variables. utils.pl has functions of common use.
Executables are called uniformly via the function utils.pl:RunTool(),
which prints out ERROR and FATAL messages, and returns a non zero if a
FATAL is encountered.

Scripts made more robust by checking the return value (of most
functions anyway) and interrupting the sequence of tools if a problem
is found.