Decouple lattice search from N-Gram search.
Though, it's worth noting that the lattice search actually does use
N-Gram models, so it's not totally clear what this gains for us :)

Oh, actually, it means this stuff can go in SphinxBase.  Which it
will, probably kind of soon, along with a lattice to FSG converter
which we would like to have for discriminative training.